backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
kerykeion==5.7.0
google-generativeai==0.3.2
python-dotenv==1.0.0
aiogram==3.3.0